Add testing
Created a testing folder that houses a vagrant file. This references the
roles above and will spin up a vagrantbox and run through the roles so
that it is possible to verify the behaviour of the playbook execution
not on a real raspberry pi.

Too may changes to list but review the diff to get an idea. Moved some
template files into template folders in their respective role folders,
updated syntax. Added var for user as vagrant uses vagrant user etc etc